waterTower/requirements.txt
Flask==1.1.1
Flask-SocketIO==4.2.1
paho-mqtt==1.5.0
flask-mqtt==1.1.1